requirements.txt
Flask
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp